// src/types/pages/node.ts
export interface INodeResult {
  NodeNo: number // Primary Key
  NodeName: string
  Validate: string // Unique
  NodeDesc: string
  NodeType: number // 1: Master, 2: Worker
  Elevator: number // 0: No, 1: Yes
  Mac: string // Unique
  Product: number // 0: Server, 1: 1 Door, 2: 2 Door, 4: 4 Door
  Model: number // 1: Advantage, 2: Professional, 3: Corporate
  Type: number // 1 ~ 512
  OemNo: number
  Oem: {
    OemNo: number
    OemName: string
  } | null
  Version: string
  Address: string
  Timezone: string
  Online: number // 0: No, 1: Yes
  PowerFaultType: number // 0: NO, 1: NC
  TamperType: number // 0: NO, 1: NC
  PowerFaultStat: number // 0: Inactive, 1: Active
  TamperStat: number // 0: Inactive, 1: Active
}

export interface ITempResult {
  TempNo: number
  NodeType: number
  Elevator: number
  Mac: string
  Product: number
  Model: number
  Type: number
  OemNo: number
  Version: string
  Address: string
  Timezone: string
}

export interface INodeFilters {
  NodeNo: string
  NodeName: string
  Mac: string
  Address: string
  Apply: boolean
}

export interface INodeRouteQueryParams {
  page: number
  NodeNo: string
  NodeName: string
  Mac: string
  Address: string
}

export interface INodeApiQueryParams extends IApiQueryParamsBase {
  NodeNo_icontains?: string
  NodeName_icontains?: string
  Mac_icontains?: string
  Address_icontains?: string
}